pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "simple-environment"
version = "0.1.1"                                   # bump so installers don't cache 0.1.0
description = "Minimal environment for Jupyter Notebook (Python 3.11) with Ollama client, for VRSecretary development."
# Chatterbox + Torch stack is really tested on 3.11; avoid 3.12 for now
requires-python = ">=3.11,<3.12"

authors = [
  { name = "Ruslan Magana Vsevolodovna" }
]

license = { text = "Apache-2.0" }

readme = "README.md"

keywords = [
  "vr",
  "ai",
  "llm",
  "ollama",
  "jupyter",
  "vrsecretary"
]

dependencies = [
  "notebook>=7.2",                                  # modern Notebook
  "ipykernel>=6.29",
  "ollama>=0.6,<1.0",                               # pin to a stable client range

  # --- Added Dependencies for ASGI/Web Server ---
  "uvicorn[standard]>=0.38.0",
  "httptools>=0.7.1",
  "pydantic-settings",                              # correct package name
  # ----------------------------------------------

  # --- TTS stack: versions aligned with chatterbox-tts 0.1.4's ecosystem ---
  "numpy==1.25.2",                                  # within chatterbox-tts range >=1.24,<1.26
  "onnx==1.19.0",                                   # works with ml-dtypes 0.5.x
  "ml-dtypes==0.5.3",                               # exposes float4_e2m1fn, fixes your crash
  "chatterbox-tts==0.1.4",                          # your TTS engine
  "requests>=2.32.0",                               # HTTP client for TTS client GUI
  "pyaudio>=0.2.14",
  "simpleaudio>=1.0.4",                             # simple cross-platform audio playback

  # NOTE: torch / torchvision / torchaudio are provided in the optional "cuda" extra
  # so environments without CUDA don't get broken installs by default.
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6",
  "pytest>=8.2",
]
lab = [
  "jupyterlab>=4.2",
]

# Install this extra when you *have* a CUDA GPU and want the Torch stack.
# CPU-only installs can ignore this.
#
# With uv + Makefile:
#   uv sync --extra cuda --extra dev --extra lab \
#     --extra-index-url https://download.pytorch.org/whl/cu126 \
#     --index-strategy unsafe-best-match
#
# With plain pip:
#   pip install ".[cuda]" --extra-index-url https://download.pytorch.org/whl/cu126
#
cuda = [
  "torch",
  "torchvision",
  "torchaudio",
]

[project.urls]
Homepage = "https://github.com/ruslanmv/VRSecretary"
Repository = "https://github.com/ruslanmv/VRSecretary"
Issues = "https://github.com/ruslanmv/VRSecretary/issues"

# 🚫 Prevent setuptools from auto-finding random local packages
[tool.setuptools]
packages = []

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["backend/gateway/tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"

# ----------------- uv build hacks -----------------
[tool.uv.extra-build-dependencies]
pkuseg = ["numpy==1.25.2"]